requirements.txt
flask==3.0.0
pycryptodome==3.19.0
bcrypt==4.0.1
gunicorn==21.2.0
PyJWT==2.3.0 